Create Narcissistic Number.c
To check whether a number is Narcissistic Number or not. n number theory, a narcissistic number in a given number base is a number that is the sum of its own digits each raised to the power of the number of digits.  The smallest example of a narcissistic number other than the trivial 1-digit numbers is 153=1^3+5^3+3^3. 	
The first few are given by 1, 2, 3, 4, 5, 6, 7, 8, 9, 153, 370, 371, 407, 1634, 8208, 9474, 54748, .............................